+++ src/attitude_check.hpp
